--- item-api/src/main/java/com/ecommerce/itemservice/async/AsyncConfig.java
@@ -1,25 +1,0 @@
-package com.ecommerce.itemservice.async;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@EnableAsync
-@Configuration
-public class AsyncConfig {
-
-    @Bean
-    public Executor asyncExecutor() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setCorePoolSize(10);
-        executor.setMaxPoolSize(10);
-        executor.setQueueCapacity(Integer.MAX_VALUE);
-        executor.setKeepAliveSeconds(30);
-        executor.setThreadNamePrefix("async-executor-");
-        executor.initialize();
-        return executor;
-    }
-}
