+++ item-api/src/main/java/com/ecommerce/itemservice/common/exception/response/ExceptionResponse.java
@@ -1,0 +1,12 @@
+package com.ecommerce.itemservice.common.exception.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class ExceptionResponse {
+
+    private final String code;
+    private final String message;
+}
