--- item-api/src/main/java/com/ecommerce/itemservice/domain/item/ItemProcessingStatus.java
@@ -1,18 +1,0 @@
-package com.ecommerce.itemservice.domain.item;
-
-public enum ItemProcessingStatus {
-
-    ITEM_NOT_FOUND,
-
-    STOCK_PRODUCTION,
-    SUCCESSFUL_PRODUCTION,
-    FAILED_PRODUCTION,
-
-    STOCK_CONSUMPTION,
-    SUCCESSFUL_CONSUMPTION,
-    FAILED_CONSUMPTION,
-    OUT_OF_STOCK,
-
-    UPDATE_SUCCESSFUL,
-    UPDATE_FAILED;
-}
