--- item-api/src/main/java/com/ecommerce/itemservice/domain/item/service/StockUpdateService.java
@@ -1,9 +1,0 @@
-package com.ecommerce.itemservice.domain.item.service;
-
-import com.ecommerce.itemservice.domain.item.ItemProcessingStatus;
-import com.ecommerce.itemservice.kafka.dto.OrderItemKafkaEvent;
-
-public interface StockUpdateService {
-
-    OrderItemKafkaEvent updateStock(OrderItemKafkaEvent orderItemKafkaEvent, ItemProcessingStatus itemProcessingStatus);
-}
