--- item-api/src/main/java/com/ecommerce/itemservice/exception/response/ExceptionResponse.java
@@ -1,12 +1,0 @@
-package com.ecommerce.itemservice.exception.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class ExceptionResponse {
-
-    private final String code;
-    private final String message;
-}
