+++ item-api/src/main/java/com/ecommerce/itemservice/item/service/StockUpdateService.java
@@ -1,0 +1,9 @@
+package com.ecommerce.itemservice.item.service;
+
+import com.ecommerce.itemservice.item.enums.ItemProcessingStatus;
+import com.ecommerce.itemservice.kafka.dto.OrderItemKafkaEvent;
+
+public interface StockUpdateService {
+
+    OrderItemKafkaEvent updateStock(OrderItemKafkaEvent orderItemKafkaEvent, ItemProcessingStatus itemProcessingStatus);
+}
